pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mobile-observe-test-framework"
version = "0.1.0"
description = "Intelligent Mobile Testing Platform - Observe, Analyze, Automate"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "Vadim Toptunov", email = "vtoptunov88@gmail.com" }
]
keywords = ["mobile", "testing", "automation", "appium", "qa", "android", "ios"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "requests>=2.31.0",
    "jsonschema>=4.20.0",
    "appium-python-client>=3.0.0",
    "selenium>=4.15.0",
    "pytest>=7.4.0",
    "pytest-bdd>=6.1.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
analyzer = [
    "tree-sitter>=0.20.0",
    "lxml>=4.9.0",
]
backend = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
]

[project.scripts]
observe = "framework.cli.main:cli"

[project.urls]
Homepage = "https://github.com/vadimtoptunov/mobile-test-recorder"
Documentation = "https://github.com/vadimtoptunov/mobile-test-recorder/docs"
Repository = "https://github.com/vadimtoptunov/mobile-test-recorder"
Issues = "https://github.com/vadimtoptunov/mobile-test-recorder/issues"

[tool.setuptools]
packages = ["framework"]

[tool.black]
line-length = 120
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.flake8]
max-line-length = 120
extend-ignore = ["E501"]
exclude = [".git", "__pycache__", ".venv", "venv", "build", "dist"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

